Moved shn command definition into cmd/shn
This required factoring out the log generator, docker client generator,
and state database connector into a separate subpackage (used
internal/). The docker client, at least, is used by the main integration
test at the root package.
